rename: s3 관련 파일 이름 수정
