+++ routes/user.route.js
@@ -23,7 +23,7 @@ import {
   getMyPenaltyPostList,
 } from "../domains/user/user.controller.js";
 import { tokenAuth } from "../middleware/token.auth.js";
-import { imageUploader } from "../middleware/image.uploader.js";
+import { imageUploader } from "../middleware/s3.js";
 
 export const userRouter = express.Router();
